refactor: feature-based 도메인 기반 폴더 구조로 변경
