--- src/basic/constants/cart.ts
@@ -1,4 +1,0 @@
-export const CART = {
-  BULK_PURCHASE_THRESHOLD: 10,
-  DEFAULT_QUANTITY_INCREMENT: 1,
-} as const;
