--- src/basic/constants/discount.ts
@@ -1,4 +1,0 @@
-export const DISCOUNT = {
-  BULK_PURCHASE_BONUS_RATE: 0.05,
-  MAX_DISCOUNT_RATE: 0.5,
-} as const;
