--- src/basic/constants/index.ts
@@ -1,9 +1,0 @@
-export * from "./calculation";
-export * from "./cart";
-export * from "./coupon";
-export * from "./defaults";
-export * from "./discount";
-export * from "./notification";
-export * from "./product";
-export * from "./search";
-export * from "./validation";
