--- src/basic/data/index.ts
@@ -1,2 +1,0 @@
-export * from "./coupon.data";
-export * from "./product.data";
