+++ src/basic/features/cart/types/cart.type.ts
@@ -1,0 +1,6 @@
+import { Product } from "@/basic/features/product/types/product";
+
+export interface CartItem {
+  product: Product;
+  quantity: number;
+}
