+++ src/basic/features/coupon/types/coupon.type.ts
@@ -1,0 +1,8 @@
+import { DiscountType } from "@/basic/features/discount/types/discount.type";
+
+export interface Coupon {
+  name: string;
+  code: string;
+  discountType: DiscountType;
+  discountValue: number;
+}
