+++ src/basic/features/discount/constants/discount.ts
@@ -1,0 +1,5 @@
+export const DISCOUNT = {
+  BULK_PURCHASE_BONUS_RATE: 0.05,
+  MAX_DISCOUNT_RATE: 0.5,
+  BULK_PURCHASE_THRESHOLD: 10,
+} as const;
