+++ src/basic/features/discount/types/discount.type.ts
@@ -1,0 +1,9 @@
+export interface Discount {
+  quantity: number;
+  rate: number;
+}
+
+export enum DiscountType {
+  AMOUNT = "amount",
+  PERCENTAGE = "percentage",
+}
