+++ src/basic/features/notification/types/notification.ts
@@ -1,0 +1,10 @@
+import { NOTIFICATION } from "@/basic/shared/constants/notification";
+
+export interface Notification {
+  id: string;
+  message: string;
+  type: NotificationType;
+}
+
+export type NotificationType =
+  (typeof NOTIFICATION.TYPES)[keyof typeof NOTIFICATION.TYPES];
