+++ src/basic/features/search/constants/search.ts
