--- src/basic/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from "./useLocalStorage";
-export * from "./useCart";
-export * from "./useProducts";
