--- src/basic/models/index.ts
@@ -1,4 +1,0 @@
-export * from "./cart.model";
-export * from "./coupon.model";
-export * from "./discount.model";
-export * from "./product.model";
