+++ src/basic/shared/constants/product.ts
