+++ src/basic/shared/hooks/useLocalStorage.ts
